changed size_t printing lines to use %lu and unsigned long cast
git-svn-id: http://svn.osgeo.org/postgis/trunk@1264 b70326c6-7e19-0410-871a-916f4a2858ee

+++ lwgeom/lwgeom_functions_basic.c
@@ -69,8 +69,9 @@ Datum LWGEOM_mem_size(PG_FUNCTION_ARGS)
 	computed_size += 4; // varlena size
 	if ( size != computed_size )
 	{
-		elog(NOTICE, "varlena size (%d) != computed size+4 (%d)",
-				size, computed_size);
+		elog(NOTICE, "varlena size (%lu) != computed size+4 (%lu)",
+				(unsigned long)size,
+				(unsigned long)computed_size);
 	}
 
 	PG_FREE_IF_COPY(geom,0);
